Changed "weight" to a Double

+++ src/Dragon.java
@@ -2,7 +2,7 @@ public class Dragon extends MagicAnimal
 {
     private int wings;
     private int scalePwr; //how hard the scales are (defense)
-    public Dragon(String name, String type, String element, int speed, int weight, int wings, int scalePwr)
+    public Dragon(String name, String type, String element, int speed, Double weight, int wings, int scalePwr)
     {
         super(name, type, element, speed, weight);
         this.wings = wings;
